+++ src/main/java/com/taxi_system/dao/CarDriverDAO.java
@@ -6,5 +6,5 @@ import com.taxi_system.db_entities.CarDriver;
  * Created by Victoria on 25.12.2017.
  */
 public interface CarDriverDAO {
-    CarDriver getById(int id);
+    CarDriver getById(long id);
 }
