+++ src/main/java/com/taxi_system/dao/DiscountDAO.java
@@ -6,6 +6,6 @@ import com.taxi_system.db_entities.Discount;
  * Created by Victoria on 25.12.2017.
  */
 public interface DiscountDAO {
-    Discount getDiscountOnSum(long sum);
-    Discount getById(int id);
+    Discount getDiscountOnSum(float sum);
+    Discount getById(long id);
 }
